refactor(app): extract genomic visualization data helpers (#346, #394)
Move the pure data-transform logic out of GenomicVisualizationTabs.vue
into genomicVisualizationData.ts: protein lollipop data assembly
(buildProteinPlotData) and exon-aware protein-to-genomic mapping
(buildExonMap, proteinToGenomic, buildGenomicVariants). The component
keeps thin computeds delegating to these helpers and stays a tab/loading
shell.

GenomicVariant (and UniProtData) now live in the helper module and are
re-exported from GenomicVisualizationTabs.vue, so existing consumers keep
importing GenomicVariant from the same path (import-preserving).

Adds genomicVisualizationData.spec.ts covering the exon-map and
protein-to-genomic seam (none existed before).

GenomicVisualizationTabs.vue 748 -> 547 lines.
